refactor apps actions
