+++ apps/chat/src/components/AppsEditor/Settings/index.tsx
@@ -83,24 +83,16 @@ export const ApplicationSettings: React.FC<Props> = ({
   schema,
   type,
 }) => {
+  const { t } = useTranslation(Translation.Chat);
+
   const router = useRouter();
+
   const dispatch = useAppDispatch();
+
   const pythonVersions = useAppSelector(
     SettingsSelectors.selectCodeEditorPythonVersions,
   );
   const modelsMap = useAppSelector(ModelsSelectors.selectModelsMap);
-  const modelFromState = applicationData
-    ? modelsMap[applicationData.reference]
-    : null;
-  const isAppDeployed = useMemo(
-    () => !!modelFromState && isApplicationDeployed(modelFromState),
-    [modelFromState],
-  );
-  const isAppDeploymentInProgress = useMemo(
-    () => !!modelFromState && isApplicationDeploymentInProgress(modelFromState),
-    [modelFromState],
-  );
-
   const previewConversationId = useAppSelector(
     ConversationsSelectors.selectPreviewConversationId,
   );
@@ -114,14 +106,24 @@ export const ApplicationSettings: React.FC<Props> = ({
     ConversationsSelectors.selectAreSelectedConversationsLoaded,
   );
   const isCodeEditorDirty = useAppSelector(CodeEditorSelectors.selectIsDirty);
-
   const theme = useAppSelector(UISelectors.selectThemeState);
-  const { t } = useTranslation(Translation.Chat);
 
   const [previewMode, setPreviewMode] = useState<PreviewMode>(
     schema?.[ApplicationTypeSchemaProperties.applicationTypeViewerUrl]
       ? PreviewMode.closed
       : PreviewMode.half,
+  );
+
+  const modelFromState = applicationData
+    ? modelsMap[applicationData.reference]
+    : null;
+  const isAppDeployed = useMemo(
+    () => !!modelFromState && isApplicationDeployed(modelFromState),
+    [modelFromState],
+  );
+  const isAppDeploymentInProgress = useMemo(
+    () => !!modelFromState && isApplicationDeploymentInProgress(modelFromState),
+    [modelFromState],
   );
 
   const getDefaultValues = useCallback(
