+++ apps/chat/src/components/Marketplace/ApplicationDetails/ApplicationFooter.tsx
@@ -1,20 +1,16 @@
 import {
   IconEdit,
   IconFileDescription,
-  IconLink,
   IconPlayerPlay,
   IconTrashX,
   IconUserShare,
   IconWorldShare,
 } from '@tabler/icons-react';
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 import classNames from 'classnames';
 
-import {
-  useMenuItemHandler,
-  useMenuItemHandlerWithTwoArgs,
-} from '@/src/hooks/useHandler';
+import { useApplicationMenuActions } from '@/src/hooks/useApplicationActions';
 import { useScreenState } from '@/src/hooks/useScreenState';
 import { useTranslation } from '@/src/hooks/useTranslation';
 
@@ -30,14 +26,12 @@ import {
 import { isMyApplication } from '@/src/utils/app/id';
 import { isEntityIdPublic } from '@/src/utils/app/publications';
 import { canWriteSharedWithMe } from '@/src/utils/app/share';
-import { getApplicationLink } from '@/src/utils/marketplace';
 
 import {
   ApplicationStatus,
   SimpleApplicationStatus,
 } from '@/src/types/applications';
 import { FeatureType, ScreenState } from '@/src/types/common';
-import { DisplayMenuItemProps } from '@/src/types/menu';
 import { DialAIEntityModel } from '@/src/types/models';
 import { Translation } from '@/src/types/translation';
 
@@ -46,12 +40,9 @@ import { AuthSelectors } from '@/src/store/auth/auth.reducers';
 import { useAppDispatch, useAppSelector } from '@/src/store/hooks';
 import { SettingsSelectors } from '@/src/store/settings/settings.reducers';
 import { ShareActions } from '@/src/store/share/share.reducers';
-import { UIActions } from '@/src/store/ui/ui.reducers';
 
 import {
   PlayerContextButtonClasses,
-  PlayerContextIconClasses,
-  PlayerContextIcons,
   StatusIcons,
 } from '@/src/constants/marketplace';
 
@@ -64,7 +55,7 @@ import { ApplicationCopyLink } from './ApplicationCopyLink';
 
 import UnpublishIcon from '@/public/images/icons/unpublish.svg';
 import IconUserUnshare from '@/public/images/icons/unshare-user.svg';
-import { Feature, PublishActions } from '@epam/ai-dial-shared';
+import { Feature } from '@epam/ai-dial-shared';
 
 const getDisabledTooltip = (entity: DialAIEntityModel, normal: string) => {
   switch (entity.functionStatus) {
@@ -83,9 +74,6 @@ interface Props {
   allVersions: DialAIEntityModel[];
   onChangeVersion: (entity: DialAIEntityModel) => void;
   onUseEntity: () => void;
-  onPublish: (entity: DialAIEntityModel, action: PublishActions) => void;
-  onEdit: (entity: DialAIEntityModel) => void;
-  onDelete: (entity: DialAIEntityModel) => void;
   onBookmarkClick: (entity: DialAIEntityModel) => void;
 }
 
@@ -93,10 +81,7 @@ export const ApplicationDetailsFooter = ({
   entity,
   allVersions,
   onChangeVersion,
-  onPublish,
   onUseEntity,
-  onEdit,
-  onDelete,
   onBookmarkClick,
 }: Props) => {
   const { t } = useTranslation(Translation.Marketplace);
@@ -120,6 +105,8 @@ export const ApplicationDetailsFooter = ({
 
   const canWrite = canWriteSharedWithMe(entity);
 
+  const { menuItems, actions } = useApplicationMenuActions({ entity });
+
   const isExecutable =
     isExecutableApp(entity) && (isMyApp || isAdmin || canWrite);
 
@@ -127,23 +114,12 @@ export const ApplicationDetailsFooter = ({
   const playerStatus = getApplicationSimpleStatus(entity);
   const isAppInDeployment = isApplicationDeploymentInProgress(entity);
 
-  const handleLogClick = useCallback(
-    (e: React.MouseEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      dispatch(ApplicationActions.getLogs(entity.id));
-      setIsOpenLogs(true);
-    },
-    [dispatch, entity.id],
-  );
-
   const handleCloseApplicationLogs = useCallback(
     () => setIsOpenLogs(false),
     [setIsOpenLogs],
   );
 
   const PlayerIcon = StatusIcons[playerStatus];
-  const PlayerContextIcon = PlayerContextIcons[playerStatus];
 
   const handleUpdateFunctionStatus = useCallback(
     (e: React.MouseEvent) => {
@@ -181,129 +157,6 @@ export const ApplicationDetailsFooter = ({
 
   const isApplicationsSharingEnabled = useAppSelector((state) =>
     SettingsSelectors.isFeatureEnabled(state, Feature.ApplicationsSharing),
-  );
-
-  const handleCopy = useCallback(
-    (e: React.MouseEvent) => {
-      e.preventDefault();
-      e.stopPropagation();
-      if (!navigator.clipboard) return;
-      const link = getApplicationLink(entity);
-      navigator.clipboard.writeText(link);
-      dispatch(UIActions.showSuccessToast(t('Link copied!')));
-    },
-    [dispatch, entity, t],
-  );
-  const handleEdit = useMenuItemHandler(onEdit, entity);
-  const handleDelete = useMenuItemHandler(onDelete, entity);
-  const handlePublish = useMenuItemHandlerWithTwoArgs(
-    onPublish,
-    entity,
-    PublishActions.ADD,
-  );
-  const handleUnpublish = useMenuItemHandlerWithTwoArgs(
-    onPublish,
-    entity,
-    PublishActions.DELETE,
-  );
-
-  const menuItems: DisplayMenuItemProps[] = useMemo(
-    () => [
-      {
-        name: t('Copy link'),
-        dataQa: 'application-copy-link',
-        display: isPublicApp && isSmallScreen,
-        Icon: IconLink,
-        onClick: handleCopy,
-      },
-      {
-        name: t(getPlayerCaption(entity)),
-        dataQa: 'status-change',
-        display: isExecutable && isCodeAppsEnabled,
-        disabled: playerStatus === SimpleApplicationStatus.UPDATING,
-        Icon: PlayerContextIcon,
-        iconClassName: PlayerContextIconClasses[playerStatus],
-        onClick: handleUpdateFunctionStatus,
-      },
-      {
-        name: t('Share'),
-        dataQa: 'share',
-        display: isMyApp && isApplicationsSharingEnabled && isSmallScreen,
-        Icon: IconUserShare,
-        onClick: handleOpenSharing,
-      },
-      {
-        name: t('Unshare'),
-        dataQa: 'unshare',
-        display: !!entity.sharedWithMe && isApplicationsSharingEnabled,
-        Icon: IconUserUnshare,
-        onClick: handleOpenUnshare,
-      },
-      {
-        name: t('Delete'),
-        dataQa: 'delete',
-        display: isMyApp,
-        disabled: isModifyDisabled,
-        Icon: IconTrashX,
-        onClick: handleDelete,
-      },
-      {
-        name: t('Publish'),
-        dataQa: 'publish',
-        display: isMyApp,
-        Icon: IconWorldShare,
-        onClick: handlePublish,
-      },
-      {
-        name: t('Unpublish'),
-        dataQa: 'unpublish',
-        display: hasPublicId,
-        Icon: UnpublishIcon,
-        onClick: handleUnpublish,
-      },
-      {
-        name: t('Edit'),
-        dataQa: 'edit',
-        display: (isMyApp || !!canWrite) && !!onEdit,
-        disabled: isAppInDeployment,
-        Icon: IconEdit,
-        onClick: handleEdit,
-      },
-      {
-        name: t('Application logs'),
-        dataQa: 'app-logs',
-        display:
-          isExecutable && playerStatus === SimpleApplicationStatus.UNDEPLOY,
-        Icon: IconFileDescription,
-        onClick: handleLogClick,
-      },
-    ],
-    [
-      t,
-      isPublicApp,
-      isSmallScreen,
-      handleCopy,
-      entity,
-      isExecutable,
-      isCodeAppsEnabled,
-      playerStatus,
-      PlayerContextIcon,
-      handleUpdateFunctionStatus,
-      isMyApp,
-      isApplicationsSharingEnabled,
-      handleOpenSharing,
-      handleOpenUnshare,
-      isModifyDisabled,
-      handleDelete,
-      handlePublish,
-      hasPublicId,
-      handleUnpublish,
-      canWrite,
-      onEdit,
-      isAppInDeployment,
-      handleEdit,
-      handleLogClick,
-    ],
   );
 
   const hasBookmark = !isMyApp && !entity.sharedWithMe;
@@ -374,7 +227,7 @@ export const ApplicationDetailsFooter = ({
                 <Tooltip tooltip={t(getDisabledTooltip(entity, 'Delete'))}>
                   <button
                     disabled={isModifyDisabled}
-                    onClick={() => onDelete(entity)}
+                    onClick={actions.handleDelete}
                     className="icon-button"
                     data-qa="application-delete"
                   >
@@ -385,13 +238,10 @@ export const ApplicationDetailsFooter = ({
               {(isMyApp || hasPublicId) && (
                 <Tooltip tooltip={hasPublicId ? t('Unpublish') : t('Publish')}>
                   <button
-                    onClick={() =>
-                      onPublish(
-                        entity,
-                        hasPublicId
-                          ? PublishActions.DELETE
-                          : PublishActions.ADD,
-                      )
+                    onClick={
+                      hasPublicId
+                        ? actions.handleUnpublish
+                        : actions.handlePublish
                     }
                     className="icon-button"
                     data-qa="application-publish"
@@ -404,11 +254,11 @@ export const ApplicationDetailsFooter = ({
                   </button>
                 </Tooltip>
               )}
-              {(isMyApp || canWrite) && (
+              {(isMyApp || canWrite || isAdmin) && (
                 <Tooltip tooltip={t('Edit')}>
                   <button
                     disabled={isAppInDeployment}
-                    onClick={() => onEdit(entity)}
+                    onClick={actions.handleEdit}
                     className="icon-button"
                     data-qa="application-edit"
                   >
@@ -420,7 +270,7 @@ export const ApplicationDetailsFooter = ({
                 playerStatus === SimpleApplicationStatus.UNDEPLOY && (
                   <Tooltip tooltip={t('Application logs')}>
                     <button
-                      onClick={handleLogClick}
+                      onClick={actions.handleOpenApplicationLogs}
                       className="icon-button"
                       data-qa="application-logs"
                     >
